various bugs, __main__ not needed anymore since django
