+++ buzzword/urls.py
@@ -1,8 +1,5 @@
 from django.contrib import admin
 from django.urls import include, path
-from explorer import __main__
-
-# from explore.views import explore
 
 urlpatterns = [
     path('', include("start.urls")),
