+++ explore/views.py
@@ -1,10 +1,7 @@
-from django.http import HttpResponse
 from django.shortcuts import render
-from explorer.__main__ import CORPORA_CONFIGS, _get_explore_layout, app
+from explorer.parts.main import populate_explorer_with_initial_data
 
 
 def explore(request, slug):
-    print('SLUG', slug)
-    print('APP', app)
-    app.layout = _get_explore_layout(slug, CORPORA_CONFIGS)
+    populate_explorer_with_initial_data(slug)
     return render(request, 'explore/explore.html')
